+++ src/services/apiClient.ts
@@ -1,11 +1,5 @@
 import { api } from "@/api/client";
-import type {
-  Reference,
-  SearchResult,
-  Enhancement,
-  BibliographicMetadataEnhancement,
-  LinkedDataEnhancement,
-} from "@/types/models";
+import type { Reference, SearchResult } from "@/types/models";
 
 export interface SearchFilters {
   page?: number;
@@ -37,28 +31,4 @@ export async function searchReferences(
 
 export async function getReference(id: string): Promise<Reference> {
   return api.get<Reference>(`/v1/references/${id}/`);
-}
-
-export function extractBibliographic(
-  reference: Reference,
-): BibliographicMetadataEnhancement | null {
-  if (!reference.enhancements) return null;
-  const matches = reference.enhancements.filter(
-    (e): e is Enhancement & { content: BibliographicMetadataEnhancement } =>
-      e.content.enhancement_type === "bibliographic",
-  );
-  if (matches.length === 0) return null;
-  return matches[matches.length - 1].content;
-}
-
-export function extractLinkedData(
-  reference: Reference,
-): LinkedDataEnhancement | null {
-  if (!reference.enhancements) return null;
-  const matches = reference.enhancements.filter(
-    (e): e is Enhancement & { content: LinkedDataEnhancement } =>
-      e.content.enhancement_type === "linked_data",
-  );
-  if (matches.length === 0) return null;
-  return matches[matches.length - 1].content;
 }
