+++ tests/services/apiClient.test.ts
@@ -1,18 +1,7 @@
 import { vi } from "vitest";
 import { api } from "@/api/client";
-import {
-  searchReferences,
-  getReference,
-  extractBibliographic,
-  extractLinkedData,
-} from "@/services/apiClient";
-import type {
-  Reference,
-  SearchResult,
-  Enhancement,
-  BibliographicMetadataEnhancement,
-  LinkedDataEnhancement,
-} from "@/types/models";
+import { searchReferences, getReference } from "@/services/apiClient";
+import type { Reference, SearchResult } from "@/types/models";
 
 vi.mock("@/api/client", () => ({
   api: {
@@ -81,128 +70,5 @@ describe("getReference", () => {
     mockedGet.mockResolvedValue(ref);
     const res = await getReference("abc-123");
     expect(res).toBe(ref);
-  });
-});
-
-function makeEnhancement(
-  content: Enhancement["content"],
-): Enhancement {
-  return {
-    id: null,
-    reference_id: "ref-1",
-    source: "test",
-    visibility: "public",
-    robot_version: null,
-    derived_from: null,
-    content,
-  };
-}
-
-function makeRef(enhancements: Enhancement[] | null): Reference {
-  return {
-    id: "ref-1",
-    visibility: "public",
-    identifiers: null,
-    enhancements,
-  };
-}
-
-describe("extractBibliographic", () => {
-  test("returns null when enhancements is null", () => {
-    expect(extractBibliographic(makeRef(null))).toBeNull();
-  });
-
-  test("returns null when no bibliographic enhancement exists", () => {
-    const ref = makeRef([
-      makeEnhancement({ enhancement_type: "abstract", process: "x", abstract: "y" }),
-    ]);
-    expect(extractBibliographic(ref)).toBeNull();
-  });
-
-  test("returns bibliographic content when present", () => {
-    const bib: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: 5,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2024,
-      publisher: null,
-      title: "Test",
-      pagination: null,
-      publication_venue: null,
-    };
-    const ref = makeRef([makeEnhancement(bib)]);
-    expect(extractBibliographic(ref)).toBe(bib);
-  });
-
-  test("returns the last bibliographic enhancement when multiple exist", () => {
-    const bib1: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: null,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2020,
-      publisher: null,
-      title: "First",
-      pagination: null,
-      publication_venue: null,
-    };
-    const bib2: BibliographicMetadataEnhancement = {
-      enhancement_type: "bibliographic",
-      authorship: null,
-      cited_by_count: null,
-      created_date: null,
-      updated_date: null,
-      publication_date: null,
-      publication_year: 2024,
-      publisher: null,
-      title: "Second",
-      pagination: null,
-      publication_venue: null,
-    };
-    const ref = makeRef([makeEnhancement(bib1), makeEnhancement(bib2)]);
-    expect(extractBibliographic(ref)).toBe(bib2);
-  });
-});
-
-describe("extractLinkedData", () => {
-  test("returns null when enhancements is null", () => {
-    expect(extractLinkedData(makeRef(null))).toBeNull();
-  });
-
-  test("returns null when no linked_data enhancement exists", () => {
-    const ref = makeRef([
-      makeEnhancement({ enhancement_type: "abstract", process: "x", abstract: "y" }),
-    ]);
-    expect(extractLinkedData(ref)).toBeNull();
-  });
-
-  test("returns linked_data content when present", () => {
-    const ld: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://example.com",
-      data: { key: "value" },
-    };
-    const ref = makeRef([makeEnhancement(ld)]);
-    expect(extractLinkedData(ref)).toBe(ld);
-  });
-
-  test("returns the last linked_data enhancement when multiple exist", () => {
-    const ld1: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://first.com",
-      data: { n: 1 },
-    };
-    const ld2: LinkedDataEnhancement = {
-      enhancement_type: "linked_data",
-      vocabulary_uri: "http://second.com",
-      data: { n: 2 },
-    };
-    const ref = makeRef([makeEnhancement(ld1), makeEnhancement(ld2)]);
-    expect(extractLinkedData(ref)).toBe(ld2);
   });
 });
